feat：项目级别的变量语法风格设置增加风险提示和风险确认 #11407

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/UserPipelineResourceImpl.kt
@@ -780,7 +780,7 @@ class UserPipelineResourceImpl @Autowired constructor(
         userId: String,
         projectId: String,
         dialect: PipelineDialectType
-    ): Result<Long> {
+    ): Result<Int> {
         return Result(
             pipelineListFacadeService.countPipelineByDialect(
                 projectId = projectId, dialect = dialect
